small refactors

+++ src/main/java/learn/Helpers.java
@@ -6,19 +6,19 @@ public class Helpers {
     // size = - n * ln(p) / (ln 2)^2
     // where:
     // double p = desired false positive
-    // long n = number of elements inserted
-    public static long calculateBitArraySize(double p, long n) {
+    // int n = number of elements inserted
+    public static int calculateBitArraySize(double p, int n) {
         double bitArraySize = - n * Math.log(p) / Math.pow(Math.log(2), 2);
-        return (long) Math.ceil(bitArraySize);
+        return (int) Math.ceil(bitArraySize);
     }
     // calculate optimal number of hash functions
     // using this equation
     // num of hash = m * ln(2) / n
     // where:
-    // long m = bit array size
-    // long n = number of elements inserted
-    public static long calculateNumOfHashFunctions(long m, long n) {
+    // int m = bit array size
+    // int n = number of elements inserted
+    public static int calculateNumOfHashFunctions(int m, int n) {
         double optimalHashes = m * Math.log(2) / n;
-        return (long) Math.floor(optimalHashes);
+        return (int) Math.floor(optimalHashes);
     }
 }
